Add exceptions handling

+++ src/pig/file_service.py
@@ -1,4 +1,5 @@
 from human_player import HumanPlayer
+import os
 
 
 class FileService:
@@ -11,6 +12,8 @@ class FileService:
 
     def load_players(self) -> list[HumanPlayer]:
         """Load the players from players.txt"""
+        if not os.path.exists(self.filename):
+          raise FileNotFoundError()
         players: list[HumanPlayer] = []
         with open(self.filename) as file:
             while True:
@@ -29,6 +32,8 @@ class FileService:
 
     def save_players(self, players: list[HumanPlayer]):
         """Save the players in players.txt"""
+        if not os.path.exists(self.filename):
+            raise FileNotFoundError()
         with open(self.filename, "w") as file:
             for player in players:
                 file.write(
@@ -38,5 +43,7 @@ class FileService:
 
     def add_player(self, player: HumanPlayer):
         """Add new Player to players.txt"""
+        if not os.path.exists(self.filename):
+          raise FileNotFoundError()
         with open(self.filename, "a") as file:
             file.write(f"{player.id}:{player.name}:\n")
